v2.03
Updated as of 06-30-2019
Completed major overhaul where file name, file location, save destination, and email recipients are no longer predetermined. They are now chosen at the time of activation. Additionally, the file is now saved with a timestamp in the destination folder and will create a historical archive; as well as the series of emails.